Moved network observers
