+++ src/main/java/org/opensearch/performanceanalyzer/commons/hwnet/NetworkInterface.java
@@ -13,11 +13,11 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.performanceanalyzer.commons.collectors.NetInterfaceSummary;
 import org.opensearch.performanceanalyzer.commons.metrics_generator.linux.LinuxIPMetricsGenerator;
+import org.opensearch.performanceanalyzer.commons.observer.ResourceObserver;
 import org.opensearch.performanceanalyzer.commons.os.metrics.NetworkMetricsCalculator;
-import org.opensearch.performanceanalyzer.commons.os.observer.ResourceObserver;
-import org.opensearch.performanceanalyzer.commons.os.observer.impl.DeviceNetworkStatsObserver;
-import org.opensearch.performanceanalyzer.commons.os.observer.impl.Ipv4Observer;
-import org.opensearch.performanceanalyzer.commons.os.observer.impl.Ipv6Observer;
+import org.opensearch.performanceanalyzer.commons.hwnet.observer.impl.DeviceNetworkStatsObserver;
+import org.opensearch.performanceanalyzer.commons.hwnet.observer.impl.Ipv4Observer;
+import org.opensearch.performanceanalyzer.commons.hwnet.observer.impl.Ipv6Observer;
 
 public class NetworkInterface {
     private static final Logger LOG = LogManager.getLogger(NetworkInterface.class);
