+++ src/main/java/org/opensearch/performanceanalyzer/commons/hwnet/observer/impl/Ipv4Observer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.performanceanalyzer.commons.os.observer.impl;
+package org.opensearch.performanceanalyzer.commons.hwnet.observer.impl;
 
 
 import java.io.BufferedReader;
@@ -14,7 +14,7 @@ import java.util.Map;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.performanceanalyzer.commons.collectors.StatsCollector;
-import org.opensearch.performanceanalyzer.commons.os.observer.ResourceObserver;
+import org.opensearch.performanceanalyzer.commons.observer.ResourceObserver;
 import org.opensearch.performanceanalyzer.commons.stats.metrics.StatExceptionCode;
 
 public class Ipv4Observer implements ResourceObserver<Long> {
